requirements.txt
asgiref==3.3.1
astroid==2.5.1
Django==3.1.7
django-appconf==1.0.4
django-compressor==2.4
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
openslide-python==1.1.2
Pillow==8.2.0
pkg-resources==0.0.0
pylint==2.7.2
python-decouple==3.4
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
Unipath==1.1
wrapt==1.12.1
